1st read_buffer test
