+++ CMakeLists.txt
@@ -82,6 +82,7 @@ if (BUILD_TESTING)
     add_unit_test(protocol          test_startup_fsm)
     add_unit_test(protocol          test_read_response_fsm)
     add_unit_test(protocol          test_check_request)
+    add_unit_test(protocol          test_next_power_of_2)
     add_unit_test(protocol          test_read_buffer)
     add_unit_test(""                test_request)
     add_unit_test(""                test_response)
